+++ collection.h
@@ -16,7 +16,7 @@ class collection
 public:
     collection(std::string filename, const catalog &catalog);
 
-    card at(uint8_t loc) const;
+    card at(uint64_t loc) const;
     uint64_t count();
 private:
     const catalog &catalog_;
